[Server] - Move model.js into the server folder
